Clean up for IMU
