+++ airbrakes/mock/MockIMU.py
@@ -1,0 +1,91 @@
+class MockIMU:
+    """
+    A mock implementation of the IMU for testing purposes. It doesn't interact with any hardware and returns data read
+    from a previous log file.
+    """
+
+    __slots__ = ("log_file_path",)
+
+    def __init__(self, log_file_path: Path, frequency: int):
+        super().__init__(frequency)
+        self.log_file_path = log_file_path
+
+        self._data_queue: multiprocessing.Queue[IMUDataPacket] = multiprocessing.Queue(MAX_QUEUE_SIZE)
+        self._running = multiprocessing.Value("b", False)
+
+        # Starts the process that fetches data from the IMU
+        self._data_fetch_process = multiprocessing.Process(
+            target=self._read_data_loop, args=(self.log_file_path, self.frequency, self.upside_down)
+        )
+
+    @property
+    def is_running(self) -> bool:
+        return super().is_running
+
+    def start(self) -> None:
+        pass
+
+    def stop(self) -> None:
+        pass
+
+    def get_imu_data_packet(self) -> IMUDataPacket | None:
+        pass
+
+    def get_imu_data_packets(self) -> collections.deque[IMUDataPacket]:
+        pass
+
+    def _read_data_loop(self, log_file_path: Path, frequency: int) -> None:
+        """
+        Reads the data from the log file and puts it into the shared queue.
+        :param log_file_path: Path to the CSV file containing IMU data.
+        :param frequency: Frequency in Hz for how often to fetch data.l
+        """
+        # Calculate the interval between readings based on frequency
+        # TODO: reading takes a little bit of time, so we should probably subtract a few milliseconds from the interval
+        interval = 1.0 / frequency
+
+        with log_file_path.open(newline="") as csvfile:
+            reader = csv.DictReader(csvfile)
+
+            for row in reader:
+                # Check if the process should stop
+                if not self._running.value:
+                    break
+
+                imu_data_packet = None
+
+                # Create the data packet based on the row
+                if row.get("scaledAccelX") is not None:
+                    imu_data_packet = RawDataPacket(int(row["timestamp"]))
+                    imu_data_packet.scaledAccelX = float(row.get("scaledAccelX"))
+                    imu_data_packet.scaledAccelY = float(row.get("scaledAccelY"))
+                    imu_data_packet.scaledAccelZ = float(row.get("scaledAccelZ"))
+                    imu_data_packet.scaledGyroX = float(row.get("scaledGyroX"))
+                    imu_data_packet.scaledGyroY = float(row.get("scaledGyroY"))
+                    imu_data_packet.scaledAccelZ = float(row.get("scaledGyroZ"))
+                elif row.get("estCompensatedAccelX") is not None:
+                    imu_data_packet = EstimatedDataPacket(int(row["timestamp"]))
+                    imu_data_packet.estCompensatedAccelX = float(row.get("estCompensatedAccelX"))
+                    imu_data_packet.estCompensatedAccelY = float(row.get("estCompensatedAccelY"))
+                    imu_data_packet.estCompensatedAccelZ = float(row.get("estCompensatedAccelZ"))
+                    imu_data_packet.estAngularRateX = float(row.get("estAngularRateX"))
+                    imu_data_packet.estAngularRateY = float(row.get("estAngularRateY"))
+                    imu_data_packet.estAngularRateZ = float(row.get("estAngularRateZ"))
+                    imu_data_packet.estOrientQuaternionW = float(row.get("estOrientQuaternionW"))
+                    imu_data_packet.estOrientQuaternionX = float(row.get("estOrientQuaternionX"))
+                    imu_data_packet.estOrientQuaternionY = float(row.get("estOrientQuaternionY"))
+                    imu_data_packet.estOrientQuaternionZ = float(row.get("estOrientQuaternionZ"))
+                    imu_data_packet.estPressureAlt = float(row.get("estPressureAlt"))
+                    imu_data_packet.estAttitudeUncertQuaternionW = float(row.get("estAttitudeUncertQuaternionW"))
+                    imu_data_packet.estAttitudeUncertQuaternionX = float(row.get("estAttitudeUncertQuaternionX"))
+                    imu_data_packet.estAttitudeUncertQuaternionY = float(row.get("estAttitudeUncertQuaternionY"))
+                    imu_data_packet.estAttitudeUncertQuaternionZ = float(row.get("estAttitudeUncertQuaternionZ"))
+
+                if imu_data_packet is None:
+                    continue
+
+                # Put the packet in the queue
+                self._data_queue.put(imu_data_packet)
+
+                # Simulate polling interval
+                time.sleep(interval)
